+++ frontend/src/app/(booking)/booking/[id]/page.tsx
@@ -1,5 +1,5 @@
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/api/auth/[...nextauth]/[...nextauth]";
 import getUserProfile from "@/libs/getUserProfile";
 import getBooking from "@/libs/getBooking";
 import CreateBookingForm from "@/components/CreateBookingForm";
