+++ frontend/src/app/mybooking/page.tsx
@@ -1,7 +1,7 @@
 import getBookings from '@/libs/getBookings';
 import {BookingItem} from "../../../interface"
 import { getServerSession } from 'next-auth';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/app/api/auth/[...nextauth]/[...nextauth]';
 import BookingHotelCard from '@/components/BookingHotelCard';
 import Link from 'next/link';
 import BookingCatalog from '@/components/ฺBookingCatalog';
