+++ frontend/src/app/page.tsx
@@ -2,7 +2,7 @@ import Banner from "@/components/Banner";
 import PromoteCard from "@/components/PromoteCard";
 import getBookings from "@/libs/getBookings";
 import { getServerSession } from "next-auth";
-import { authOptions } from "./api/auth/[...nextauth]/route";
+import { authOptions } from "./api/auth/[...nextauth]/[...nextauth]";
 import getUserProfile from "@/libs/getUserProfile";
 export default async function Home() {
